--- pages/app.vue
@@ -1,12 +1,0 @@
-<template>
-  <div class="flex flex-col min-h-screen">
-    <Header />
-    <NuxtPage />
-    <Footer />
-  </div>
-</template>
-
-<script setup>
-import Footer from '~/components/Footer.vue'
-import Header from '~/components/Header.vue' // komt later
-</script>
